refactor(crdt): centralize computed outcomes
