+++ packages/shared-server/rallar-system/crdt/mutation/compute-crdt-mutation.ts
@@ -1,34 +1,21 @@
 import {
   byteLengthOfRallarCrdtJson,
   hashRallarCrdtUpdateEnvelope,
-  type RallarCrdtAppendRejected,
-  type RallarCrdtAppendResult,
   type RallarCrdtDocumentMetadata,
+  type RallarCrdtTrustedAppendMetadata,
   validateRallarCrdtUpdateEnvelope,
 } from '@shared/crdt/mod.ts';
 import type {
   CrdtAppendCommand,
-  CrdtCanonicalSnapshotEnvelope,
-  CrdtMutationCommand,
   CrdtMutationComputed,
-  CrdtMutationComputedRejected,
-  CrdtMutationComputedReplay,
-  CrdtMutationComputedWrite,
   CrdtMutationAttemptFacts,
   CrdtMutationRead,
 } from './crdt-mutation-contracts.ts';
-import { toAppendOutbox, toCrdtAuditOutbox } from './create-crdt-mutation-outbox.ts';
 import {
-  appendRejectionReason,
-  isAppendRejectionRetryable,
-  toAppendRejectionCode,
-} from './crdt-append-rejection.ts';
-import {
-  toAcceptedAdminResultDetails,
-  toCrdtMutationResult,
-  toFallbackReplayAppend,
-  toRejectedAdminResultDetails,
-} from './create-crdt-mutation-result.ts';
+  computeCrdtAcceptedAppendOutcome,
+  computeCrdtRejectedOutcome,
+  computeCrdtReplayOutcome,
+} from './compute-crdt-mutation-outcome.ts';
 import {
   computeCrdtErase,
   computeCrdtLifecycleUpdate,
@@ -37,32 +24,6 @@ import {
 } from './compute-crdt-administration.ts';
 
 export interface ComputeCrdtMutationInput extends CrdtMutationAttemptFacts {
-  readonly serviceId: string;
-}
-
-interface CrdtMutationRejectedInput {
-  readonly command: CrdtMutationCommand;
-  readonly read: CrdtMutationRead;
-  readonly code: string;
-  readonly serviceId: string;
-}
-
-interface CrdtMutationComputedBaseInput<TDocument extends RallarCrdtDocumentMetadata | null> {
-  readonly command: CrdtMutationCommand;
-  readonly read: CrdtMutationRead;
-  readonly document: TDocument;
-  readonly update: CrdtAppendCommand['update'] | null;
-  readonly append: CrdtMutationComputed['append'];
-  readonly snapshot: CrdtCanonicalSnapshotEnvelope | null;
-  readonly outboxEntries: CrdtMutationComputed['outboxEntries'];
-  readonly result: CrdtMutationComputed['result'];
-}
-
-interface CrdtMutationWriteComputedInput {
-  readonly command: Exclude<CrdtMutationCommand, CrdtAppendCommand>;
-  readonly read: CrdtMutationRead;
-  readonly document: RallarCrdtDocumentMetadata;
-  readonly snapshot: CrdtCanonicalSnapshotEnvelope | null;
   readonly serviceId: string;
 }
 
@@ -76,7 +37,7 @@ interface NextAppendDocumentInput {
 export function computeCrdtMutation(input: ComputeCrdtMutationInput): CrdtMutationComputed {
   const { command, read, serviceId } = input;
   if (!read.authorized) {
-    return rejected({
+    return computeCrdtRejectedOutcome({
       command,
       read,
       code: read.authorizationCode === 'allowed' ? 'authorization-denied' : read.authorizationCode,
@@ -84,7 +45,7 @@ export function computeCrdtMutation(input: ComputeCrdtMutationInput): CrdtMutati
     });
   }
   if (!read.featureDecision.allowed) {
-    return rejected({ command, read, code: 'feature-disabled', serviceId });
+    return computeCrdtRejectedOutcome({ command, read, code: 'feature-disabled', serviceId });
   }
   switch (command.operation) {
     case 'append':
@@ -107,7 +68,7 @@ function computeCrdtAppend(input: ComputeCrdtMutationInput): CrdtMutationCompute
   }
   const validation = validateRallarCrdtUpdateEnvelope(command.update);
   if (!validation.valid) {
-    return rejected({ command, read, code: 'invalid-update', serviceId });
+    return computeCrdtRejectedOutcome({ command, read, code: 'invalid-update', serviceId });
   }
   const candidateHash = hashRallarCrdtUpdateEnvelope(command.update);
   if (read.existingUpdate) {
@@ -116,11 +77,11 @@ function computeCrdtAppend(input: ComputeCrdtMutationInput): CrdtMutationCompute
         ? null
         : 'duplicate-hash-mismatch';
     return code === null
-      ? replay({ command, read, serviceId })
-      : rejected({ command, read, code, serviceId });
+      ? computeCrdtReplayOutcome({ command, read, serviceId })
+      : computeCrdtRejectedOutcome({ command, read, code, serviceId });
   }
   if (read.document && read.document.lifecycle !== 'active') {
-    return rejected({
+    return computeCrdtRejectedOutcome({
       command,
       read,
       code: `document-${read.document.lifecycle}`,
@@ -133,26 +94,26 @@ function computeCrdtAppend(input: ComputeCrdtMutationInput): CrdtMutationCompute
     quota?.maxUpdateCount !== undefined &&
     (read.document?.updateCount ?? 0) >= quota.maxUpdateCount
   ) {
-    return rejected({ command, read, code: 'quota-exceeded', serviceId });
+    return computeCrdtRejectedOutcome({ command, read, code: 'quota-exceeded', serviceId });
   }
   if (quota?.maxUpdateBytes !== undefined && updateBytes > quota.maxUpdateBytes) {
-    return rejected({ command, read, code: 'update-too-large', serviceId });
+    return computeCrdtRejectedOutcome({ command, read, code: 'update-too-large', serviceId });
   }
   if (
     quota?.maxDocumentBytes !== undefined &&
     (read.document?.storedUpdateBytes ?? 0) + read.storedSnapshotBytes + updateBytes >
       quota.maxDocumentBytes
   ) {
-    return rejected({ command, read, code: 'quota-exceeded', serviceId });
+    return computeCrdtRejectedOutcome({ command, read, code: 'quota-exceeded', serviceId });
   }
   if (
     quota?.maxUpdatesPerMinutePerActor !== undefined &&
     read.actorUpdatesInWindow >= quota.maxUpdatesPerMinutePerActor
   ) {
-    return rejected({ command, read, code: 'rate-limited', serviceId });
+    return computeCrdtRejectedOutcome({ command, read, code: 'rate-limited', serviceId });
   }
   const appendSequence = (read.document?.lastAppendSequence ?? 0) + 1;
-  const append = {
+  const append: RallarCrdtTrustedAppendMetadata = {
     appendSequence,
     acceptedAtEpochMs: command.capturedAtEpochMs,
     actorId: command.actor.actorId,
@@ -161,35 +122,9 @@ function computeCrdtAppend(input: ComputeCrdtMutationInput): CrdtMutationCompute
     serverId: command.actor.serverId,
     authorizationScope: command.authorizationScope,
     acceptedUpdateHash: candidateHash,
-  } as const;
+  };
   const document = nextAppendDocument({ command, read, appendSequence, updateBytes });
-  const appendResult: RallarCrdtAppendResult = {
-    status: 'accepted',
-    update: command.update,
-    append,
-    document,
-  };
-  const result = toCrdtMutationResult({
-    command,
-    status: 'accepted',
-    document,
-    appendSequence,
-    code: null,
-    details: { appendResult },
-  });
-  return {
-    ...createCrdtMutationComputedBase({
-      command,
-      read,
-      document,
-      update: command.update,
-      append,
-      snapshot: null,
-      outboxEntries: toAppendOutbox({ command, response: appendResult, serviceId, fanout: true }),
-      result,
-    }),
-    outcome: 'write',
-  };
+  return computeCrdtAcceptedAppendOutcome({ command, read, document, append, serviceId });
 }
 
 function nextAppendDocument(input: NextAppendDocumentInput): RallarCrdtDocumentMetadata {
@@ -222,164 +157,4 @@ function nextAppendDocument(input: NextAppendDocumentInput): RallarCrdtDocumentM
     quota: read.featureDecision.policy?.quota ?? null,
     projectionIds: [],
   };
-}
-
-export function writeComputed(input: CrdtMutationWriteComputedInput): CrdtMutationComputedWrite {
-  const { command, read, document, snapshot, serviceId } = input;
-  const resultDetails = toAcceptedAdminResultDetails({ command, read, document, snapshot });
-  const mutationResult = toCrdtMutationResult({
-    command,
-    status: 'accepted',
-    document,
-    appendSequence: document.lastAppendSequence,
-    code: null,
-    details: resultDetails,
-  });
-  const auditEvent =
-    mutationResult.operation === 'erase' && mutationResult.status === 'accepted'
-      ? mutationResult.auditEvent
-      : null;
-  return {
-    ...createCrdtMutationComputedBase({
-      command,
-      read,
-      document,
-      update: null,
-      append: null,
-      snapshot,
-      outboxEntries: auditEvent ? [toCrdtAuditOutbox(auditEvent, command, serviceId)] : [],
-      result: mutationResult,
-    }),
-    outcome: 'write',
-  };
-}
-
-interface ReplayCrdtMutationInput {
-  readonly command: CrdtAppendCommand;
-  readonly read: CrdtMutationRead;
-  readonly serviceId: string;
-}
-
-function replay(input: ReplayCrdtMutationInput): CrdtMutationComputedReplay {
-  const { command, read, serviceId } = input;
-  const append = read.existingAppend ?? toFallbackReplayAppend(command, read.document);
-  const appendResult: RallarCrdtAppendResult = read.document
-    ? { status: 'duplicate', update: command.update, append, document: read.document }
-    : rejectionResult(command, null, 'storage-failed');
-  const response = toCrdtMutationResult({
-    command,
-    status: 'replay',
-    document: read.document,
-    appendSequence: append.appendSequence,
-    code: null,
-    details: { appendResult },
-  });
-  return {
-    ...createCrdtMutationComputedBase({
-      command,
-      read,
-      document: read.document,
-      update: command.update,
-      append,
-      snapshot: null,
-      outboxEntries: toAppendOutbox({ command, response: appendResult, serviceId, fanout: false }),
-      result: response,
-    }),
-    outcome: 'replay',
-  };
-}
-
-export function rejected(input: CrdtMutationRejectedInput): CrdtMutationComputedRejected {
-  const { command, read, code, serviceId } = input;
-  const appendResult =
-    command.operation === 'append' ? rejectionResult(command, read.document, code) : undefined;
-  const response = toCrdtMutationResult({
-    command,
-    status: 'rejected',
-    document: read.document,
-    appendSequence: null,
-    code,
-    details: appendResult
-      ? { appendResult }
-      : toRejectedAdminResultDetails(command as Exclude<CrdtMutationCommand, CrdtAppendCommand>),
-  });
-  const outboxEntries =
-    command.operation === 'append' &&
-    appendResult !== undefined &&
-    read.authorized &&
-    !code.startsWith('authorization-') &&
-    !code.startsWith('authentication-')
-      ? toAppendOutbox({ command, response: appendResult, serviceId, fanout: false })
-      : [];
-  return {
-    ...createCrdtMutationComputedBase({
-      command,
-      read,
-      document: read.document,
-      update: command.operation === 'append' ? command.update : null,
-      append: null,
-      snapshot: null,
-      outboxEntries,
-      result: response,
-    }),
-    outcome: 'rejected',
-    code,
-  };
-}
-
-function createCrdtMutationComputedBase<TDocument extends RallarCrdtDocumentMetadata | null>(
-  input: CrdtMutationComputedBaseInput<TDocument>,
-): {
-  readonly command: CrdtMutationCommand;
-  readonly read: CrdtMutationRead;
-  readonly operation: CrdtMutationCommand['operation'];
-  readonly commandId: string;
-  readonly commandHash: string;
-  readonly documentKey: string;
-  readonly expectedDocumentRevision: number | 'absent';
-  readonly expectedDocumentLifecycle: CrdtMutationComputed['expectedDocumentLifecycle'];
-  readonly expectedAppendSequence: number | 'absent';
-  readonly document: TDocument;
-  readonly update: CrdtAppendCommand['update'] | null;
-  readonly append: CrdtMutationComputed['append'];
-  readonly snapshot: CrdtCanonicalSnapshotEnvelope | null;
-  readonly outboxEntries: CrdtMutationComputed['outboxEntries'];
-  readonly result: CrdtMutationComputed['result'];
-} {
-  const { command, read, document, update, append, snapshot, outboxEntries, result } = input;
-  return {
-    command,
-    read,
-    operation: command.operation,
-    commandId: command.commandId,
-    commandHash: command.commandHash,
-    documentKey: command.documentKey,
-    expectedDocumentRevision: read.document?.documentRevision ?? 'absent',
-    expectedDocumentLifecycle: read.document?.lifecycle ?? 'absent',
-    expectedAppendSequence: read.document?.lastAppendSequence ?? 'absent',
-    document,
-    update,
-    append,
-    snapshot,
-    outboxEntries,
-    result,
-  };
-}
-
-function rejectionResult(
-  command: CrdtAppendCommand,
-  document: RallarCrdtDocumentMetadata | null,
-  code: string,
-): RallarCrdtAppendRejected {
-  const rejectionCode = toAppendRejectionCode(code);
-  const rejection = {
-    status: 'rejected',
-    update: command.update,
-    code: rejectionCode,
-    reason: appendRejectionReason(rejectionCode),
-    ...(document ? { document } : {}),
-  } as const;
-  return isAppendRejectionRetryable(rejectionCode)
-    ? { ...rejection, code: rejectionCode, retryable: true }
-    : { ...rejection, code: rejectionCode, retryable: false };
 }
